# Improve superseeding performance by scaling the number of superseeded pieces to a peer's upload speed