--- a/simulation/test_super_seeding.cpp
+++ b/simulation/test_super_seeding.cpp
@@ -10,6 +10,7 @@ see LICENSE file.
 #include "setup_swarm.hpp"
 #include "libtorrent/add_torrent_params.hpp"
 #include "libtorrent/settings_pack.hpp"
+#include "settings.hpp"
 #include "test.hpp"
 
 using namespace lt;
@@ -29,6 +30,31 @@ TORRENT_TEST(super_seeding)
 		// terminate
 		, [](int, lt::session&) -> bool
 		{ return true; });
+}
+
+TORRENT_TEST(superseed_no_redundant_have)
+{
+	sim::default_config network_cfg;
+	sim::simulation sim{network_cfg};
+
+	lt::add_torrent_params default_add_torrent;
+
+	lt::settings_pack default_settings = settings();
+	default_settings.set_bool(settings_pack::send_redundant_have, false);
+	default_settings.set_bool(settings_pack::close_redundant_connections, false);
+
+	setup_swarm(5, swarm_test::upload, sim, default_settings, default_add_torrent
+		// add session
+		, [](lt::settings_pack& pack) {}
+		// add torrent
+		, [](lt::add_torrent_params& params) {
+			params.flags |= torrent_flags::super_seeding;
+		}
+		// on alert
+		, [](lt::alert const*, lt::session&) {}
+		// terminate
+		, [](int, lt::session&) -> bool
+		{ return false; });
 }
 
 #if TORRENT_ABI_VERSION == 1