--- a/src/settings_pack.cpp
+++ b/src/settings_pack.cpp
@@ -356,7 +356,8 @@ namespace libtorrent {
 		SET(dht_max_infohashes_sample_count, 20, nullptr),
 		SET(max_piece_count, 0x200000, nullptr),
 		SET(min_websocket_announce_interval, 1 * 60, nullptr),
-		SET(webtorrent_connection_timeout, 2 * 60, nullptr)
+		SET(webtorrent_connection_timeout, 2 * 60, nullptr),
+		SET(superseed_scale_time, 6, nullptr)
 	}});
 
 #undef SET