… bindings/c/include/libtorrent_settings.h