To-do:

- [ ] Possibly improve handling of peers that filter redundant `HAVE` messges
- [x] Simulations
- [ ] Unit tests

---

In order to improve transfer speed a peer can queue up a number of blocks for download. However when superseeding, peers will only be able to queue data from the one or two pieces we've said that we have, in some cases limiting the achievable upload speed. This is more of an issue with a faster upload capacity, higher latency, and for torrents with a smaller piece size.

For instance, for a torrent with an 8MB piece size, a peer can potentially request up to 16MB (depending on `max_allowed_in_request_queue` and their clients behavior) of data, as it believes that we only have those 2 pieces. In the case of a 256KB piece size, the peer could only request up to a max of 512KB though. Note these are maximums, and deplete as the pieces download until we tell the peer we `HAVE` another piece.

In order to have upload speeds to peers on superseeded torrents up to par with normally seeded torrents, we would need to increase the number of pieces we make available to each peer from the current hard limit of 2.

For testing I reworked the superseed behavior to make use of a new option `max_superseed_bytes`. The number of pieces to offer a peer when superseeding would then be this divided by the torrent's piece size. These are the results I got from a test torrent with 256KB sized pieces with a single tcp peer with ~150ms latency and ~2.5MB/s bandwidth:

| `max_superseed_bytes` | Equiv number of 256KB pieces offered| Observed speed |
|---|---|---|
|524288|2 (current behavior)|~0.35MB/s|
|1048576|4|~0.4MB/s|
|2097152|8|~0.8MB/s|
|4194304|16|~1.2MB/s|
|8388608|32|~1.8MB/s|
|10485760|40|~2.1MB/s|
|12582912|48|~2.3MB/s|
|16777216|64|~2.3MB/s|
|x|All - Standard seeding|~2.3MB/s|

While this seems to be enough to improve performance, it is still a bit blunt and requires manually configuring to suit the upload speed. Also there is no point, for instance, announcing 8MB (32x256KB) worth of pieces to a peer that's only downloading from us at 50KB/s. With a large number of slower peers this may impact superseeding efficiency.

A better solution, and the one in this PR, is to dynamically increase/decrease the number of pieces superseeded to a peer depending on our upload speed to them, similar to the way `request_queue_time` is used for sizing the download request queue. Having added a `superseed_scale_time` setting I reran the previous test, and the number of pieces superseeding automatically increased to 53 as the speed increased and stabilized, which is about the sweet spot in this case and on par with standard seeding. That was with `superseed_scale_time=6` - ie. offering as many pieces at a time as we expect the peer to download from us in 6s.

Running some further tests involving higher latency (~550ms) it looks like a higher default value may be worthwhile, so maybe 8-10s or so. I don't think a higher value is wasteful, as actual requests is still limited by `max_allowed_in_request_queue` - it'll just give the peer a larger range of pieces to choose from. I wouldn't imagine it's high enough to effect superseeding efficiency at this level either.

I'm reasonably confident in the logic behind this PR - I don't think superseeding efficiency would be impacted, and the performance improvements seem easy enough to demonstrate. It does stray from the [BEP](https://www.bittorrent.org/beps/bep_0016.html) somewhat, though I see that 'strict' superseeding as the BEP decsribes has been deprecated since 1.2.6, so hopefully trying to 'improve' on the BEP in this way isn't an issue. 

As I'm still relatively new to c++ stuff I can't rule out amateur errors in the code. That said, I'm pleased with how it's been working so far, and I think its time to get some feedback on it and whether this behavior, or something similar, might be worth implementing.