@thrnz 2021-07-26
1. I have a number of otherwise seedless torrents with only a limited upload speed. I've found that using superseeding makes more efficient use of what I've got to spare compared to standard seeding.

2. Sure. I'll rebase this to master. 

3. I'll look into this and see if I can sort something. It might above my level at this stage but keen to learn.