@1265578519 2021-07-26
你可以在制作种子的时候设置较小的区块就可以解决这个问题，更改这个些代码可能会导致其他客户端不适配引起BUG，对方客户端区块未下载完成无法尝试上传