@arvidn 2021-07-26
> You can solve this problem by setting a smaller block when making the seed. Changing these codes may cause other clients to not adapt and cause BUG, and the other client’s block cannot be uploaded without downloading.

The piece size is immutable, you can't change it without changing the info-hash