@arvidn 2021-07-27
Another, pre-existing, issue is that this logic can easily be gamed. You just need to send a HAVE message back to get another super seeded piece to open up. You can do that until the "right" piece opens up.

I think that mitigation would be to keep track of bytes we upload, and not always trust HAVE messages. It's a tricky balance though, since there may be other seeds sending data too.