@thrnz 2021-07-27
I've rebased to master and changed m_superseed_pieces from `vector` to `set`.

`target_superseed_pieces()` also now takes into account the number of pieces the peer has remaining, and makes sure it never returns a target higher than that. That should prevent any weirdness as the upload nears completion.

Dealing with peers that filter redundant `HAVE` messages remains unchanged at this stage.

I did some experimenting with different values of `superseed_scale_time` with high latencies and got the following. This was with a torrent with 256KB pieces and a single TCP peer.

Peer latency | `superseed_scale_time` | Observed speed
--- | --- | ---
300ms | 2s | 0.1-0.2MB/s
300ms | 3s | 1.0-1.3MB/s (slow to increase)
300ms | 4s | 1.6-1.8MB/s
300ms | 6s | 2.1-2.3MB/s
300ms | 8s | 2.1-2.3MB/s
300ms | normal seeding | 2.1-2.3MB/s
600ms | 3s | 0.1MB/s
600ms | 4s | 0.1MB/s
600ms | 5s | 1.3-1.7MB/s (slow to increase)
600ms | 6s | 2.0-2.3MB/s (slow to increase)
600ms | 8s | 2.0-2.3MB/s
600ms | normal seeding | 2.1-2.3MB/s

From that I guess 6s should be enough for most cases to reach parity with standard seeding.