@thrnz 2021-09-10
The current `super_seeding` and `strict_super_seeding` tests in `simulation/test_super_seeding.cpp` appear to always pass even if super seeding is intentionally broken. If I'm reading things correctly, it appears that both tests are finishing before the downloads have a chance to complete/fail.

I've changed them so they now wait until the downloaders either successfully become seeds or timeout/fail before the tests terminate.