@thrnz 2021-12-31
I've done some further testing and got the following `bmon` graphs showing network utilization for various values of `superseed_scale_time`. These were done while super-seeding multiple torrents with varying piece sizes to a mix of both TCP and uTP peers of various latencies, so is more 'real world' than the previous tests which were only to a single peer.


```
                           superseed_scale_time=0
                  (ie. single piece offered/current behavior)
     MiB                      (TX Bytes/second)
    2.63 ................................|..|........................
    2.19 |.........|..||.................|..|...|....................
    1.75 |..|...|..||.||.........|||...|.|..||..|...||...|.......||..
    1.31 |..|..|||.||.||.||.|.||||||||||||.|||.||...|||||||..||.|||||
    0.88 ||||||||||||||||||.||||||||||||||||||.||||||||||||.|||||||||
    0.44 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
         1   5   10   15   20   25   30   35   40   45   50   55   60

                            superseed_scale_time=2
     MiB                      (TX Bytes/second)
    2.77 ...|......................................|.................
    2.31 ...||..||........|.|.||..|...|.|......|..|||...|......|...||
    1.84 .|.||..||...|.|..|||.||..|..||.|.|||..|..||||.||.|||.|||.|||
    1.38 |||||..|||||||||.||||||.||||||||.|||.|||||||||||||||||||||||
    0.92 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.46 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
         1   5   10   15   20   25   30   35   40   45   50   55   60

                            superseed_scale_time=4
     MiB                      (TX Bytes/second)
    2.88 ||||.|||.|||...|..||||..|.|...||..|....||.....|....|.|||||.|
    2.40 |||||||||||||||||.|||||||||||||||||||.|||||||.|....|||||||||
    1.92 |||||||||||||||||.||||||||||||||||||||||||||||||||||||||||||
    1.44 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.96 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.48 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
         1   5   10   15   20   25   30   35   40   45   50   55   60

                            superseed_scale_time=6
     MiB                      (TX Bytes/second)
    2.88 ||||||||||||||....||||||||||||||||.|||.||||||||||||||||||.||
    2.40 ||||||||||||||||..||||||||||||||||||||||||||||||||||||||||||
    1.92 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    1.44 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.96 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.48 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
         1   5   10   15   20   25   30   35   40   45   50   55   60

                            superseed_scale_time=8
     MiB                      (TX Bytes/second)
    2.88 |||||||||||||||||||||||||||||||||||||.|||||||.||||||||||.|||
    2.40 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    1.92 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    1.44 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.96 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
    0.48 ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
         1   5   10   15   20   25   30   35   40   45   50   55   60
```